move all images to a folder for better organization
