+++ app/src/MyStocks/styles.module.scss
@@ -48,7 +48,6 @@
         text-decoration: none;
         color: darkblue;
         font-size: smaller;
-        font-weight: 1000;
 
         &:hover {
           color: rgb(133, 133, 133);
