+++ app/src/Nav/styles.module.scss
@@ -25,7 +25,7 @@ nav {
       padding: 5px;
 
       &:hover {
-        color: yellow;
+        color: lightpink;
         border-bottom: #70e2ac 2px solid;
         text-decoration: none;
       }
