+++ app/src/global.scss
@@ -10,10 +10,12 @@ html {
   box-sizing: border-box;
 }
 
-*,
-*:before,
-*:after {
-  box-sizing: inherit;
+* {
+  font-weight: 700;
+  &:before,
+  &:after {
+    box-sizing: inherit;
+  }
 }
 
 main {
@@ -23,12 +25,15 @@ main {
 h1 {
   font-size: 2rem;
   margin: 1.5rem 0 0.5rem;
+  font-weight: 700;
 }
 
 h2 {
   font-size: 1.5rem;
+  font-weight: 700;
 }
 
 h3 {
   font-size: 1rem;
+  font-weight: 700;
 }
