+++ app/src/images/discoverIdeas.png
