+++ app/src/images/stockChart.png
